separated concerns: DTO for login and another for registration
