+++ src/main/java/com/arola/wc/controllers/ContactFormController.java
@@ -4,13 +4,14 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import com.arola.wc.api.Visitors;
+import com.arola.wc.api.RegistrationData;
 
 @Controller
 public class ContactFormController {
 	
 	@RequestMapping("/showContactForm")
-	public String showContactForm(@ModelAttribute("user") Visitors visitor) {		
+	public String showContactForm(@ModelAttribute("user") 
+	RegistrationData visitor) {		
 		
 		return "contact-form";
 	}
@@ -18,7 +19,7 @@ public class ContactFormController {
 	// handler for processing form
 	
 	@RequestMapping("/processContactForm")
-	public String processContactForm(@ModelAttribute("visitor") Visitors visitor) { 
+	public String processContactForm(@ModelAttribute("visitor") RegistrationData visitor) { 
 		
 		// access the attribute names in the view page to get the inputs and display them
