+++ src/main/java/com/arola/wc/controllers/LoginController.java
@@ -7,13 +7,13 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import com.arola.wc.api.Visitors;
+import com.arola.wc.api.LoginData;
 
 @Controller
 public class LoginController {
 	
 	@RequestMapping("/showLoginForm")
-	public String showLoginForm(@ModelAttribute("user") Visitors user) {		
+	public String showLoginForm(@ModelAttribute("user") LoginData user) {		
 		
 		return "login";
 	}
@@ -22,10 +22,9 @@ public class LoginController {
 	
 	@RequestMapping("/processLoginForm")
 	public String processContactForm( @Valid @ModelAttribute("user") 
-	Visitors user, BindingResult result) { 
+	LoginData user, BindingResult result) { 
 		
 		if(result.hasErrors()) {
-			
 			return "login"; // stay on login page.
 		}
